requirements.txt
aiofiles
appdirs
asgiref
async-timeout
asyncio
attrs
backports.entry-points-selectable
byte
CacheControl
certifi
cffi
chardet
charset-normalizer
click
colorama
cryptography
py-cord
py-cord[voice]
py-cord[speed]
distlib
distro
Django
filelock
geographiclib
geojson
geopy
google
greenlet
idna
image
more-itertools
msgpack
multidict
odict
ordered-set
Pillow
pip
platformdirs
plumber
ply
progress
pycparser
PyNaCl
pyOpenSSL
pyowm
pyparsing
PySocks
python-dateutil
pytz
requests
resolvelib
retrying
setuptools
six
soupsieve
sqlparse
toml
tomli
typing-extensions
virtualenv
webencodings
Werkzeug
xmltodict
yarl
youtube-dl
zope.component
zope.deprecation
zope.event
zope.hookable
zope.interface
zope.lifecycleevent
asyncpraw
python-dotenv